Generate app showcase content

This function generates a showcase panel for Shiny apps.

app_showcase(app_url, img_url, title, subtitle, label = NULL, drop = NULL,
  height = 200, min_width = 300, max_width = 400, col_width = 4,
  img_padding = NULL)

Arguments

app_url
character vector of app urls.

img_url
character vector of image urls, same length as app_url.

title
character, titles displayed on image links during mouseover.

subtitle
character, subtitles displayed on image link during mouseover.

label
character or NULL, optional fixed label place above an image panel. See details.

drop
character vector of any apps to exclude from showcase. Useful for programmed app lists that may include self.

height
numeric, height in pixels.

min_width
numeric, minimum width in pixels.

max_width
numeric, maximum width in pixels.

col_width
integer, column width number for row, 1 through 12, defaults to 4.

img_padding
optional image padding, e.q., "10px". Defaults to NULL. See details. If length is one, is repeated for length of app_url.

Value

a shiny fluidRow containing organized and stylized app image links for reference.

Details

Despite being named for this apputils context, app_showcase can be applied to other types of content.

The image padding argument is useful when multiple source images have different amounts of margin space embedded in them and you need them to match each other better, or when you need padding around an image that includes no margin around it. For image padding, use a vector of length one or length equal to the length of app_url.

If label is not NULL, it must be a vector with a label for every image link. To force empty labels among non-empty labels, use "". E.g., label = c("A", "", "C"). label can be styled by using html, e.g., label = h4("Top Title").

Examples

#not run