The apputils package contains common utility functions, settings and references for development use across multiple Shiny apps.

apputils has a shinydashboard focus. apputils is a member package in the satellite sector of the SNAPverse. It supports other satellites in the verse, including maputils and snaputils.

Details

Package functionality and areas of support covered by apputils include:

Overrides of shinydashboard::valueBox, shinydashbaord::infoBox and shiny::icon that support the use of local thumbnails images.

Stat boxes: special type of value or info boxes for common statistics using a collection of icons provided by the package.

Adjusted CSS styles and integration with packages like rintrojs and shinytoastr for interactive tours and toast messages.

Functions for including app information widgets such as citations, contact info, frequently asked questions and more.

Encapsulation of working with data frames in server.R in specific contexts and use cases to simplify code.

Wrappers around specific use cases for Leaflet maps, data tables, and general plotting in apps.

Helper functions for dynmaic reports.