Add footer to Shiny dashboard sidebar

Add a foot to a Shiny dashboard sidebar that includes a logo image link and a label.

dashboard_footer(href, src, label = "", width = "100%", height = "160px",
  italic = TRUE, bold = TRUE,
  style = "text-align:center; align: center; padding: 0px; margin: 0px;")

Arguments

href
url, will open in new window.

src
image url.

label
character, should be short.

width
width of logo.

height
height of footer, defaults to "160px".

italic
logical, for label text.

bold
logical, for label text.

style
character, option style string for logo and label. See function for defaults.

Value

an html footer tag.

Details

style pertains to the logo and label whereas width is for the logo specifically and only height pertains directly to the footer tag.

Examples

#not run