Accessor for apputils icons

Access path to local apputils stat box icon image files.

statIcon(id, theme = "white")

Arguments

id
character, id to be matched to a stat box icon.

theme
character, "white" (default) or "black".

Value

character string specifying the local image file location in the apputils package.

Details

This accessor function provides convenient access to local package resources. Specifically, given an ID, it returns the resource path to a stat box icon png file. This is typically used in a Shiny app that has apputils loaded. Available options include:

normal

Normal distribution density curve

min

Normal curve with minimum line and statistic notation

max

As above, for the maximum

mean

As above, for the mean

sd

As above, with two lines marking +/- one standard deviation

median

As above, for the median

iqr

Box plot with interquartile range marked and labeled

b0

Beta zero hat for linear models, scatter plot with linear model intercept highlight

b1

Beta one hat for linear models, scatter plot with linear model slope highlight

r2

R-squared, scatter plot with correlation highlight

pvalue

Normal distribution curve with lower tail highlight

inc

Increasing trend line with delta change symbol

dec

As above but decreasing

pctinc

As above but with percentage symbol added

pctdec

As above but decreasing and with percentage symbol added

barinc

Increasing consecutive bars with arrow change indicators

bardec

As above but decreasing

Examples

statIcon("normal")

#> [1] "resources/images/stat_icon_normal_dist_white.png"