Initialize apputils

Call this function once near the top of the Shiny UI definition.

use_apputils(use_rintrojs = FALSE, use_shinytoastr = FALSE)

Arguments

use_rintrojs
logical, also loads rintrojs package UI components via introjsUI.

use_shinytoastr
logical, also loads shinytoastr package UI components via useToastr.

Value

HTML tags placed in the head of the app UI HTML file.

Details

Calling use_apputils loads package css into the head of a Shiny app. Optionally calls introjsUI() and/or useToastr(), in which case the Shiny app must have these packages loaded.

Examples

#not run