Genrate CSS for different colored value boxes.

Generate CSS for background color and text color of value boxes based on hex color input.

valueBoxColorsCSS(bg = "#333333", col = "#FFFFFF")

Arguments

bg
character, hex color of value box background.

col
character, hex color of value box text.

Value

a character string wrapped in shiny::HTML.

Details

Colors must be in hex color format, e.g., "#FFFFFF". This function generates CSS to be added to a Shiny app. New value box elements are named by appending the hex color to .bg- without the #. When using the valueBox override function in apputils, make sure to specify a hex color "name" that is a character string without the #, e.g., "FFFFFF". The only available colors that will work with valueBox in an app are those that have their CSS generated and loaded in the app as well.

Examples

#not run